+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+- **Breaking**: Bump `opendal` to 0.54
+
+## Removed
+- **Breaking**: Remove `OpendalStore`
+  - Use `AsyncOpendalStore` with the `AsyncToSyncStorageAdapter` instead
+
 ## [0.7.2] - 2025-05-16
 
 ## Changed